feat: support global snap align
