+++ packages/vstory-core/src/utils/layout.ts
@@ -1,6 +1,7 @@
 import type { IRect } from '@visactor/vrender-core';
 import type { IWidgetData } from '../interface/dsl/dsl';
-import type { ICharacter } from '../interface/character';
+import type { ICharacter, ILayoutLine } from '../interface/character';
+import type { IAABBBounds } from '@visactor/vutils';
 
 export interface ILayoutAttribute {
   x: number;
@@ -37,4 +38,64 @@ export function getLayoutFromWidget(w: Partial<IWidgetData> | IRect, character: 
     angle: (w as any).angle ?? 0,
     anchor: [x + width / 2, y + height / 2].map(item => (isFinite(item) ? item : 0)) as [number, number]
   };
+}
+
+export function getLayoutLine(b: IAABBBounds, opt: any, orient: 'x' | 'y' | 'xy' = 'xy') {
+  const result: ILayoutLine[] = [];
+  if (orient === 'y' || orient === 'xy') {
+    const commonInY: Omit<ILayoutLine, 'value' | 'type'> = {
+      orient: 'y',
+      start: b.x1,
+      end: b.x1 + b.width(),
+      bounds: b.clone(),
+      ...opt
+    };
+    // top
+    result.push({
+      value: b.y1,
+      type: 'start',
+      ...commonInY
+    });
+    // bottom
+    result.push({
+      value: b.y2,
+      type: 'end',
+      ...commonInY
+    });
+    // middle
+    result.push({
+      value: (b.y1 + b.y2) * 0.5,
+      type: 'middle',
+      ...commonInY
+    });
+  }
+
+  if (orient === 'x' || orient === 'xy') {
+    const commonInX: Omit<ILayoutLine, 'value' | 'type'> = {
+      orient: 'x',
+      start: b.y1,
+      end: b.y2,
+      bounds: b.clone(),
+      ...opt
+    };
+    // left
+    result.push({
+      value: b.x1,
+      type: 'start',
+      ...commonInX
+    });
+    // right
+    result.push({
+      value: b.x2,
+      type: 'end',
+      ...commonInX
+    });
+    // middle
+    result.push({
+      value: (b.x1 + b.x2) * 0.5,
+      type: 'middle',
+      ...commonInX
+    });
+  }
+  return result;
 }
